[robotInterface] Partially re-write xml parsing
The tag "<file name="...">" is now replaced by 3 different tags:
* <params file="...">
* <devices file="...">
* <actions file="...">
The root tag in the included file must be the same of the tag including
it, but without the file attribute.

This change was necessary in order to allow 2 different devices in the
same included file (for example for ems it is necessary to instantiate
two devices, one for motion control and one for the skin.

Also the examples were updated.

svn path=/trunk/iCub/; revision=21779