fixed mis-behavior with 4dc boards: added a 1ms delay between the transmitted eth frames. The mis-behavior was due to the fact that the ems cannot process udp packets with gaps of tens of microsec (the isr is not that fast to execute and many packets fill the dma buffers).
added a sequence number in udp-can-gateway packets (to be used for debug).

removed a udp message which forces a reboot if ems is in application.

changed timings with the aim to be able to catch the can boards in bootloader or in application.

The above changes are compatible with current version of the eUpdater (v2.2 of 25 sept 2013).

But if one wants to catch the can boards in bl or appl must use a newer version of eUpdater (the soon-to-be-released v2.3 is ok).



svn path=/trunk/iCub/; revision=26330