Coman lib: robolli is now an external depencency
In order to find the library, the envinronment library COMAN_DIR has to be set pointing to installation folder of the robolli library.

The FindRobolli.cmake has been added to the conf folder as helper to find the library.

the coman_handler is a singleton wrapping the library and used by the other modules like comanMotionControl and so on.

Old version of the library stored inside the repo has been removed.

std=c++0x compilation flag has been added to coman make targets (and only for them) in order to compile correctly.

svn path=/trunk/iCub/; revision=27814