icub-firmware-shared: now modules that need firmware repository checks for it
iCubInterface and canLoader lib will be skipped, an INFO message will be printed
to warn the user that they will not be compiled, but cmake won't stop since they
are only used on the robot or by expert users.

Module canBus* and embObj* will print an error message telling the user to disable
them if the library is not found. CMake will stop in this case. Since those modules
are needed while compiling the iCub&robotInterface the error is clear when compiling
stuff for the robot.

The error message reports the location of the github repository.