emObjSkin: Correctly initialize data vector to be used with new firmware version.
The new firmware does not broadcast anymore the data msgs for triangle
physically not connected/missing (while before MTB was sending a default value).
This was resulting in a error at higher level (skinManager).
Now we read the 'noLoad' value from the config file and use it to initialize vector.