robotInterface: check if calibrator device is valid before calling park action.
It coulld happens that the open of a device fails before opening the calibrator, therefore
the calibrator is not created at all. Then robotInterface closes calling all the actions, one
of them is park and therefore it will be called on a non-existing device.
This patch make this call safe.