camCalibWithPose: Cleanup and use callback for torso encoders
Using a normal port does not ensure that we are reading the right
envelope. See also robotology/yarp#610 about the reason why we need the
callback here.

robotology/yarp#623 was apparently introduced with
d44481c4b9fc13f21d1fc5dceb17d4e9d1f8afe4, when the torso was added,
because we never noticed anything similar in the past, therefore this
commit might be related.