[iCubLisboa01]: right hand thumb opposition (j8) calibration
Tested on the real robot.

"Encoders" and "Zeros" values were determined with the standard Excel procedure using the measurements:
* thumb at 0 deg => raw encoder value "Vmin" = 3000
* thumb at 90 deg => raw encoder value "Vmax" = 2000

Then, because in our case Vmax<Vmin (poles in the magnet are inverted with respect to the common situation, and it is not an option to spin the magnet because it is glued), the Excel formula for "calibration1" was wrong for us. We calculated the correct value like this:
* "calibration1" = "Vmin" - (10 deg expressed in raw units)
* then we had to switch the signs of the three PID control terms "kp", "kd", "ki"
* then we adjusted the value of "kp" to compensate overshooting

TODO: update iCub_Calibration_V1_3_1_iCubLisboa01.xls with (i) values of Vmin, Vmax; (ii) macro that works for our magnet orientation.

See #195, cc @lorejam