[iCubLisboa01]: right hand thumb opposition (j8) calibration
Tested on the real robot.

These calibration values were obtained with the new replacement magnet in May 2016 and the standard Excel procedure using the following raw measurements:

* thumb at 0 deg (finger open) => raw encoder value "Vmin" = 2000
* thumb at 90 deg (perpendicular to palm) => raw encoder value "Vmax" = 3500

This time the orientation of the magnet is standard, so Vmin<Vmax, so the PID gains have the correct signs according to specs. Just for reference, here was the previously rejected Pull Request: https://github.com/robotology/icub-main/pull/307

See #195

TODO: update the Excel sheet iCub_Calibration_V1_3_1_iCubLisboa01.xls with the new values of Vmin, Vmax (not sure which is the policy or good practice there, perhaps @randaz81 or @maggia80 know)