[skinDynLib] Add forceTorqueEstimateConfidence
Add a new integer attribute (that is correctly transmitted over the network)
that represents the confidence on the force/torque estimate trasmitted
with the skinContact . The actual meaning is implementation defined, but
the idea that more precise estimation (for example estimation using calibrated
skin) should fill this parameter with a greater integer w.r.t. rough estimations
of the contact force.