BUGFIX: FirmwareUpdater now correctly manages the tx of calibrationDataSNxxx.dat file also in cfw2 drivers.
the problem: if we attempt to program the calibration data of the strain in the case of cfw2 driver,
the can frames are sent as long bursts and the boards loses some of them because of its short rx buffer.

the solution: added small delays to split the long bursts.