The calibration RMS is saved in a file within the pictures folder
Changed the calibration file to be saved inside the folder with the images, avoiding to be overwritten by the following callibrations

removing redundat rms file writting

now the RMS for each individual calibration is also saved in the file. The order or RMSs is: left, right, stereo

moved the images dir creation right before starting the calibration. As such empty dir will not be created every time the module starts

created a new rpc instruction to make the calibration with the images in a folder

finished cfi instruction, changed the update(in|ex)trinsic functions to get the output file as an argument so they can be resused in the cfi command

Changed so a new folder is created to save results every time the start rpc command is get so the user does not need to close and restart the module to perform a new calibration. Moved the image folder input to the startCalib() method instead of the stereoCalibThread() constructor.

Now the RMS values are saved in the stereo(mon0)calibrationRun() functions instead of stereo(mono)calibration(). The return value of stereo(mono)calibration function were changed to return the rms.

fixed bug on file selection which make two consecutive runs of cfi to fail. Added cfi_shuffle option

Added module option for defining the output calibration file name. Returning the calibration file to the calibration folder to keep the calibration flow complient with the documentation. The File now is copied inside the images folder for backup.

change NULL by nullptr and sprintf by snprintf

getting some lost updates

fixed bug on appending rms results

Added the cfi_reshuffle option, which reshuffles the images to create more calibration sets. Fixed a bug where random was always using a fixed seed.

added documentation and authors

correting merging bug

retabbing