Fix implementation of embObjMultipleFTsensors::getNrOfTemperatureSensors()
Due to how the embObjMultipleFTsensors::getNrOfTemperatureSensors() function
was implemented and how the internal buffers of embObjMultipleFTsensors were
initialized, the embObjMultipleFTsensors::getNrOfTemperatureSensors() method
was return 0 until an actual temperature measure was reported. This commit
fixes this by using to compute getNrOfTemperatureSensors a buffer that has
the correct shape even if no temperature measure has been read.